Change Structure of projectData to JS Object
Change projectData back into a JS Object (per project reuiqrements) and assign temperature, date, and feelings as properties.

There were some difficulties in trying to assign properties via the entire req.body as it was, so properties are individually assigned as of now. Figuring a way to refactor the code may be a consideration in the future.